M5T1 Refactoring a module with utilities
